added: service interface

+++ bookStore/src/main/java/com/book/store/service/productservice/ProductService.java
@@ -5,22 +5,41 @@ import com.book.store.repository.productrepository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
-public class ProductService {
+public class ProductService implements ProductServiceInterface {
+
     @Autowired
-  private  ProductRepository productRepository;
-    public Product addProduct(Product product){
+    private ProductRepository productRepository;
+
+    @Override
+    public Product addProduct(Product product) {
         return productRepository.save(product);
-    }
-    public Iterable<Product> getAllProducts(){
-        return productRepository.findAll();
-    }
-    public Product updateProduct(Long id, Product product){
-        product.setId(id);
-        return productRepository.save(product);
-    }
-    public void deleteProduct(Long id){
-        productRepository.deleteById(id);
     }
 
+    @Override
+    public Iterable<Product> getAllProducts() {
+        return productRepository.findAll();
+    }
+
+    @Override
+    public Product updateProduct(Long id, Product product) {
+        Optional<Product> existingProduct = productRepository.findById(id);
+        if (existingProduct.isPresent()) {
+            product.setId(id);
+            return productRepository.save(product);
+        } else {
+            throw new RuntimeException("Product not found with id " + id);
+        }
+    }
+
+    @Override
+    public void deleteProduct(Long id) {
+        if (productRepository.existsById(id)) {
+            productRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Product not found with id " + id);
+        }
+    }
 }
